Add separate module for buffer element types
