+++ Sprocket/Graphics/mesh.cpp
@@ -5,7 +5,6 @@
 #include <assimp/Importer.hpp>
 #include <assimp/postprocess.h>
 #include <assimp/scene.h>
-#include <glad/glad.h>
 
 #include <cassert>
 #include <ranges>
@@ -218,20 +217,6 @@ void LoadSkeleton(
 
 }
 
-void static_vertex::set_buffer_attributes()
-{
-    for (int index : std::views::iota(0, 5)) {
-        glEnableVertexAttribArray(index);
-        glVertexAttribDivisor(index, 0);
-    } 
-
-    glVertexAttribPointer(0, 3, GL_FLOAT, GL_FALSE, sizeof(static_vertex), (void*)offsetof(static_vertex, position));
-    glVertexAttribPointer(1, 2, GL_FLOAT, GL_FALSE, sizeof(static_vertex), (void*)offsetof(static_vertex, textureCoords));
-    glVertexAttribPointer(2, 3, GL_FLOAT, GL_FALSE, sizeof(static_vertex), (void*)offsetof(static_vertex, normal));
-    glVertexAttribPointer(3, 3, GL_FLOAT, GL_FALSE, sizeof(static_vertex), (void*)offsetof(static_vertex, tangent));
-    glVertexAttribPointer(4, 3, GL_FLOAT, GL_FALSE, sizeof(static_vertex), (void*)offsetof(static_vertex, bitangent));
-}
-
 static_mesh_data static_mesh_data::load(const std::string& file)
 {
     Assimp::Importer importer;
@@ -378,22 +363,12 @@ void static_mesh::bind() const
 
 
 animated_mesh::animated_mesh(const animated_mesh_data& data)
-    : d_vertex_buffer(0)
-    , d_index_buffer(0)
-    , d_vertex_count(data.indices.size())
+    : d_vertices()
+    , d_indices()
     , d_skeleton(data.skeleton)
 {
-    glCreateBuffers(1, &d_vertex_buffer);
-    glNamedBufferData(d_vertex_buffer, sizeof(animated_vertex) * data.vertices.size(), data.vertices.data(), GL_STATIC_DRAW);
-
-    glCreateBuffers(1, &d_index_buffer);
-    glNamedBufferData(d_index_buffer, sizeof(std::uint32_t) * data.indices.size(), data.indices.data(), GL_STATIC_DRAW);
-}
-
-animated_mesh::~animated_mesh()
-{
-    glDeleteBuffers(1, &d_vertex_buffer);
-    glDeleteBuffers(1, &d_index_buffer);
+    d_vertices.set_data(data.vertices);
+    d_indices.set_data(data.indices);
 }
 
 std::unique_ptr<animated_mesh> animated_mesh::from_data(const animated_mesh_data& data)
@@ -408,23 +383,8 @@ std::unique_ptr<animated_mesh> animated_mesh::from_file(const std::string& file)
 
 void animated_mesh::bind() const
 {
-    glBindBuffer(GL_ARRAY_BUFFER, d_vertex_buffer);
-    glBindBuffer(GL_ELEMENT_ARRAY_BUFFER, d_index_buffer);
-    
-    for (int index : std::views::iota(0, 7)) {
-        glEnableVertexAttribArray(index);
-        glVertexAttribDivisor(index, 0);
-    }
-
-    glVertexAttribPointer(0, 3, GL_FLOAT, GL_FALSE, sizeof(animated_vertex), (void*)offsetof(animated_vertex, position));
-    glVertexAttribPointer(1, 2, GL_FLOAT, GL_FALSE, sizeof(animated_vertex), (void*)offsetof(animated_vertex, textureCoords));
-    glVertexAttribPointer(2, 3, GL_FLOAT, GL_FALSE, sizeof(animated_vertex), (void*)offsetof(animated_vertex, normal));
-    glVertexAttribPointer(3, 3, GL_FLOAT, GL_FALSE, sizeof(animated_vertex), (void*)offsetof(animated_vertex, tangent));
-    glVertexAttribPointer(4, 3, GL_FLOAT, GL_FALSE, sizeof(animated_vertex), (void*)offsetof(animated_vertex, bitangent));
-    glVertexAttribIPointer(5, 4, GL_INT, sizeof(animated_vertex), (void*)offsetof(animated_vertex, boneIndices));
-    glVertexAttribPointer(6, 4, GL_FLOAT, GL_FALSE, sizeof(animated_vertex), (void*)offsetof(animated_vertex, boneWeights));
-
-    glBindBuffer(GL_ARRAY_BUFFER, 0);
+    d_vertices.bind();
+    d_indices.bind();
 }
 
 std::vector<glm::mat4> animated_mesh::get_pose(const std::string& name, float time) const
